requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
torch==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu
